Add files via upload
Added the ability for the brands dropdown menu to interact successfully with the website scraping functions.
Added a user input field that interacts with website scraping functions.
Bestbuy and Acer are currently giving me issues. Don't have time to fix them right now.

All functions currently print out to the console and do not display on the application window.